+++ NEWS.md
@@ -1,4 +1,10 @@
-# RcppCWB 0.5.5.9001-.9003
+# RcppCWB 0.6.0.9001ff
+
+* New function `region_to_strucs()` to get minimumum and maximum struc of 
+s-attribute within region provided. Works also for nested s-attributes.
+
+
+# RcppCWB 0.6.0
 
 * Rcpp wrappers for Corpus Library (CL) functions are exposed directly and  
 can be used in C++ functions imported using `Rcpp::sourceCpp()` or